Modularize code into different folders
