+++ app/src/main/java/edu/washington/cs/ubicomplab/rdt_reader/views/RDTCameraView.java
@@ -6,7 +6,7 @@
  * of a BSD-style license that can be found in the LICENSE file.
  */
 
-package edu.washington.cs.ubicomplab.rdt_reader;
+package edu.washington.cs.ubicomplab.rdt_reader.views;
 
 import android.content.Context;
 import android.graphics.Rect;
