+++ app/src/main/res/layout/activity_image_quality.xml
@@ -13,12 +13,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.washington.cs.ubicomplab.rdt_reader.ImageQualityActivity">
+    tools:context="edu.washington.cs.ubicomplab.rdt_reader.activities.ImageQualityActivity">
 
-    <edu.washington.cs.ubicomplab.rdt_reader.ImageQualityView
+    <edu.washington.cs.ubicomplab.rdt_reader.views.ImageQualityView
         android:id="@+id/imageQualityView"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
-    </edu.washington.cs.ubicomplab.rdt_reader.ImageQualityView>
+    </edu.washington.cs.ubicomplab.rdt_reader.views.ImageQualityView>
 
 </LinearLayout>
