+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.washington.cs.ubicomplab.rdt_reader.MainActivity">
+    tools:context="edu.washington.cs.ubicomplab.rdt_reader.activities.MainActivity">
 
     <Button
         android:id="@+id/expdateButton"
