+++ app/src/main/res/layout/image_quality_view.xml
@@ -13,9 +13,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="edu.washington.cs.ubicomplab.rdt_reader.ImageQualityActivity">
+    tools:context="edu.washington.cs.ubicomplab.rdt_reader.activities.ImageQualityActivity">
 
-    <edu.washington.cs.ubicomplab.rdt_reader.AutoFitTextureView
+    <edu.washington.cs.ubicomplab.rdt_reader.views.AutoFitTextureView
 
         android:id="@+id/texture"
         android:layout_width="match_parent"
@@ -40,7 +40,7 @@
         android:textSize="20sp"
         app:layout_constraintStart_toStartOf="parent" />
 
-    <edu.washington.cs.ubicomplab.rdt_reader.ViewportUsingBitmap
+    <edu.washington.cs.ubicomplab.rdt_reader.views.ViewportUsingBitmap
         android:id="@+id/img_quality_check_viewport"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
